chore(ui): migrate IntroBox to ts
